Added a command line option to control the generation of proxy molecules
The visualization command line option (-z) accepts a numeric parameter
which is a collection of binary flags and values. This commit uses the
0x08 bit to signal whether to export proxy molecules or not. The default
is off (0) which means do not export proxy molecules.

History:

The early MCellR code only generated proxy molecules in the viz files.
All of the molecules (in those versions) appeared identical with the
only detectable difference being between volume and surface molecules.
The proxy molecules really represented the entire complex. Of course,
CellBlender can only display what's in the files, so there were only
two kinds of molecules in those early versions: volume and surface.

As MCell began to export additional viz blocks based on the molecules
within each complex, CellBlender began to display them as well along
with the proxies. Eventually, CellBlender added specific options and
defaults to hide the proxies (even though they were still in the file).

At this point, the visualization capabilities of MCellR have matured
to the point that the proxies are no longer needed. In fact they tend
to double the file sizes. Furthermore, they add complexity for the new
custom visualization code that has to filter them out of the files.

For all of these reasons, this version of MCell now defaults to NOT
exporting the proxies. However, since it's not clear if this feature
might still be useful, the VIZ_PROXY_OUTPUT flag is recognized in the
"-z" option in the fourth bit position (0x08). When this bit is set,
MCell should produce the same output as it did before this commit.