Removed trailing comma from molecule instance list in JSON Output
Python's JSON importer does not appear to tolerate trailing spaces in a
list (as Python itself does). So when exporting JSON viz format, MCell
must keep track of whether a given line is the last in the list or not.

The previous logic didn't do that (or didn't do it correctly). This
version only prints the previous comma before it is about to print the
next line. Even though this logic is slightly more complicated, it can
better handle cases where the instance loop terminates for unusual or
complicated reasons.