Added code to map graph pattern entries into table indexes
The graph pattern strings are used as keys into the molecule structure
dictionary. However, the structure data itself is saved as a list for
fast indexed lookup. The list of structures is produced by traversing
the dictionary which may be in any order. This is solved by creating a
table that maps graph pattern IDs into their location in the output
table. This is implemented in the integer array "gp_entry_for_id". The
array is built as the graph patterns are added to the dictionary, and
it is used as a second level of indirection when writing the instances.

Note that this was only done for the JSON output and not the "DUMP"
output format.