Added simple parsing of states from the Nauty string into bnglviz file.
The bnglviz file format has included an empty list for the states. This
commit replaces that empty list with a string containing the entire list
of states (with tilde prefixes for each state) or an empty string (if
there are no states).

Note that there is memory allocation that is not yet being freed. This
should be done as soon as possible.