Added an explicit "perl" prefix to the BNG2.pl command.
The previous version had simply run the BNG2.pl as if it were an
executable of some sort. In UNIX-based systems, such a text file is
opened and examined for a "#!" line such as this one:

   #!/usr/bin/perl

This is interpreted to mean that the script should be opened with the
application "/usr/bin/perl".

This notation, however, is not native Windows behavior. So in Windows,
the more proper approach is to explicitly name the application along
with the script to be run:

  "perl BNG2.pl"

This commit makes that change in mdlr2mdl.py. It has been tested to work
in Linux with the assumption that "perl" exists in the user's path.