Fixed assignment of orientation to reactions in the form S + V -> S.V
MCell3R/BNG assumes that orientation is set, i.e., S' + V' -> S'.V',
so for reactions with complexes containing more than a single molecule,
we must set orientation and this then affects rxn rate which is
multiplied by two. The FCERI model depends on this.

However, in MCell3 reactions S1 + V -> S2, the default orientation is
S1; + V; -> S2; and so the rate is not mutliplied by two.