Reorganize geometry generated by data_model_to_pymcell
This change re-organizes the geometry generated by data_model_to_pymcell.
Previously all geometry was put into a single python file.  Python seems to
have performance problem is the size of the geometry file is too large and it
can take many minutes to import the geometry of the neuropil reconstruction
model, for example.  However if the neuropil geometry is separated into
separate files, or even just one object per file, the import of all the
geometry takes just 20 seconds or so.

So, here we now create the geometry as a geometry python package where the
objects are put in separate files (i.e. separate python modules) with one
object per file.  The separate files are referenced in the __init__.py file
of the geometry package.