Skip "FILE:" observables in data_model_to_pymcell
"FILE:" observables can be specified in the Plot Output Settings of CellBlender
This is a convenient way to include externally generated files in one's plots.
And in MCell4 one can do postprocessing of data and generate such files as a
customization of an MCell4 Python model.  But the data_model_to_pymcell
converter was trying convert these during export.  This commit makes the
converter ignore "FILE:" observables.