Fix leakage due to placement of volume products near reflective boundaries
Volume products of reactions near reflective boundaries could sometimes be
placed on the opposite side of the boundary.  This could occur when the moving
reactant is transparent to the boundary but the target reactant is reflective
to the boundary.  The solution is to detect that this is the case and only if
this is the case then place the product at the location of the target molecule.
This maintains reversibility in the general case when the reaction is not near
such a boundary condition.